migrates to app router, adds some new features
