+++ app/sign-up/page.js
@@ -1,0 +1,7 @@
+'use client';
+
+import SignUpForm from '../../components/SignUpForm.js';
+
+export default function SignUp() {
+  return <SignUpForm />
+}
