+++ components/NewPollForm.js
@@ -1,6 +1,6 @@
 import clsx from 'clsx';
 import { useState } from 'react';
-import { useRouter } from 'next/router';
+import { useRouter } from 'next/navigation';
 import Link from 'next/link';
 
 import Polls from '../services/polls';
