+++ db/index.js
@@ -1,10 +1,7 @@
 import { drizzle } from 'drizzle-orm/libsql';
-import { createClient } from "@libsql/client";
 
-const client = createClient({
-  url: process.env.TURSO_DB_URL,
-  authToken: process.env.TURSO_DB_AUTH_TOKEN,
-});
-
-export const db = drizzle(client);
+export const db = drizzle({ connection: {
+  url: process.env.TURSO_DB_URL, 
+  authToken: process.env.TURSO_DB_AUTH_TOKEN 
+}});
