+++ drizzle.config.js
@@ -1,7 +1,6 @@
 export default {
   schema: "./db/schema/*",
-  dialect: "sqlite",
-  driver: "turso",
+  dialect: "turso",
   dbCredentials: {
     url: process.env.TURSO_DB_URL,
     authToken: process.env.TURSO_DB_AUTH_TOKEN,
