+++ package.json
@@ -10,27 +10,27 @@
     "db:migrate": "ts-node ./db/migrate.js"
   },
   "devDependencies": {
-    "@headlessui/react": "^1.7.7",
+    "@headlessui/react": "^2.0.0",
     "@heroicons/react": "^2.0.13",
     "@tailwindcss/forms": "^0.5.3",
     "autoprefixer": "^10.4.2",
     "date-fns": "^2.29.3",
-    "drizzle-kit": "^0.21.2",
+    "drizzle-kit": "^0.30.6",
     "eslint": "8.7.0",
-    "eslint-config-next": "^13.1.1",
+    "eslint-config-next": "^13.4.0",
     "lodash": "^4.17.21",
     "postcss": "^8.4.5",
     "tailwindcss": "^3.0.15",
     "ts-node": "^10.9.2"
   },
   "dependencies": {
-    "@libsql/client": "^0.5.6",
+    "@libsql/client": "^0.15.2",
     "@paralleldrive/cuid2": "^2.2.2",
-    "@react-email/components": "^0.0.18",
+    "@react-email/components": "^0.0.36",
     "axios": "^1.3.2",
     "clsx": "^1.2.1",
-    "drizzle-orm": "^0.30.2",
-    "next": "^13.1.1",
+    "drizzle-orm": "^0.41.0",
+    "next": "^15.2.4",
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
     "resend": "^3.2.0"
