--- pages/sign-up.js
@@ -1,5 +1,0 @@
-import SignUpForm from '../components/SignUpForm.js';
-
-export default function SignUp() {
-  return <SignUpForm />
-}
